requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
et-xmlfile==1.1.0
idna==3.3
jsonpickle==2.0.0
openpyxl==3.0.9
python-Levenshtein==0.12.2
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
